requirements_linux.txt
alembic==1.3.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
decorator==4.4.1
Flask==1.1.1
Flask-Migrate==2.5.2
flask-oidc==1.4.0
Flask-Session==0.3.1
Flask-SQLAlchemy==2.4.1
future==0.16.0
httplib2==0.14.0
idna==2.8
infinity==1.4
intervals==0.8.1
itsdangerous==1.1.0
Jinja2==2.10.3
Mako==1.1.0
MarkupSafe==1.1.1
mock==2.0.0
oauth2client==4.1.3
pbr==5.4.4
pyasn1==0.4.8
pyasn1-modules==0.2.7
PyJWT==1.7.1
python-dateutil==2.8.1
python-editor==1.0.4
python-ldap==3.2.0
requests==2.22.0
rsa==4.0
sh==1.12.14
six==1.13.0
SQLAlchemy==1.3.11
SQLAlchemy-Utils==0.35.0
urllib3==1.25.7
validators==0.14.1
Werkzeug==0.16.0
wmi-client-wrapper-py3==2018.3
WTForms==2.2.1
WTForms-Alchemy==0.16.9
WTForms-Components==0.10.4